requirements.txt
six
pyyaml
ansible-lint
requests
PyGithub
